Code I used to test different components
Hardware train controller was ECE 1140 program, but I am using it here to test the display that I am connecting to the Arduino to make sure it works correctly; this program can be used to verify that the display functions correctly. The accelerometer testing code was added to test the correct functionality of the accelerometer.